Code documentation and cleaning.
