+++ Assets/Scripts/NewGame/Entities/GameEntity.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b7f915c7113deae4188e00d425c0098d
-timeCreated: 1467472455
+guid: e5ac5ce8e9bd1eb4cb1442c9548542c0
+timeCreated: 1467545519
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
