+++ Assets/Scripts/NewGame/Interfaces/ColorTintable.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 4087fc272f769bb43b64d69cfe58de05
+timeCreated: 1467545275
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
