+++ Assets/Scripts/NewGame/Interfaces/Eliminable.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 50aa0963436d3444ea4131803404e452
+timeCreated: 1467545258
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
