+++ Assets/Scripts/NewGame/Interfaces/GameElements.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 95083044aed3a0f48839d1e332a17bb9
+folderAsset: yes
+timeCreated: 1467545491
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
