+++ Assets/Scripts/NewGame/Interfaces/ObjectCreatable.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: eb2b8c1d6d8155544b5c4f93213cceff
+timeCreated: 1467545290
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
