+++ Assets/Scripts/NewGame/Interfaces/Positionable.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: adf88101062a89743a2d81d28a22c377
+timeCreated: 1467545208
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
