+++ Assets/Scripts/NewGame/Interfaces/Speedable.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 0077b6053c8ebee48b48f9f2ea1bfb67
+timeCreated: 1467545243
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
